split rule to different files
