--- eslint/index.ts
@@ -1,9 +1,0 @@
-// const definedDependencies = require("./rules/definedDependencies/definedDependencies");
-
-// module.exports = {
-//     rules: {
-//         'definedDependencies': {
-//             create:definedDependencies,
-//         }
-//     }
-// };
