--- eslint/rules/index.ts
@@ -1,7 +1,0 @@
-// import definedDependencies from "./definedDependencies/definedDependencies";
-
-// const allRules = {
-//     "defined-dependencies": definedDependencies
-//     };
-
-// export default allRules;
