--- eslint/utils/createRule.ts
@@ -1,5 +1,0 @@
-import {ESLintUtils} from "@typescript-eslint/utils";
-
-export const createRule = ESLintUtils.RuleCreator(
-    (ruleName) => ruleName
-);
